Safari-609.1.11.1/LayoutTests/storage/websql/open-database-creation-callback-isolated-world-expected.txt

CONSOLE MESSAGE: line 6: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: line 42: Web SQL is deprecated. Please use IndexedDB instead.
ALERT: undefined
ALERT: PASS: document.body.bar visible in a callback created in this world.
This test tests that the openDatabase() creation callback is called in the right world.
